requirements.txt
fastapi
uvicorn
python-multipart
jinja2
Pillow
numpy
sqlalchemy
requests
